force computation of bw for small transfers

+++ test/zoidfs-microbenchmarks/zoidfs-tile-io.c
@@ -690,7 +690,7 @@ void report_results(MPI_Comm comm,
 	/* calculate the I/O bandwidth while we still have the max. time;
 	 * avoid floating point exceptions
 	 */
-	double d_total_size = total_size / (int64_t) 1048576;
+	double d_total_size = total_size / 1048576.0;
 	if (max_time != 0.0)
 	    io_bw = ((double) d_total_size) / max_time;
 	else
